+++ src/views/team-info.js
@@ -1,5 +1,5 @@
 import React, {Component} from 'react'
-import GraphSet from '../components/graph-set'
+import GraphSet from '../components/chart-set'
 import TeamInfoParser from '../data-parsing/team-info-parser'
 import TeamInfoConfig from '../data-parsing/configs/team-info-config'
 import MainMenu from './main-menu'
